@@ -1,15 +1,15 @@
 [tool.poetry]
 name = "mealie"
-version = "0.1.0"
+version = "0.3.0"
 description = "A Recipe Manager"
 authors = ["Hayden <hay-kot@pm.me>"]
 license = "MIT"
 
 [tool.poetry.scripts]
-start = "app:app"
+start = "mealie.app:main"
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.9"
 aiofiles = "0.5.0"
 aniso8601 = "7.0.0"
 appdirs = "1.4.4"
@@ -25,14 +25,37 @@ PyYAML = "^5.3.1"
 extruct = "^0.12.0"
 scrape-schema-recipe = "^0.1.3"
 python-multipart = "^0.0.5"
+fastapi-camelcase = "^1.0.2"
+bcrypt = "^3.2.0"
+python-jose = "^3.2.0"
+passlib = "^1.7.4"
+lxml = "4.6.2"
 
 [tool.poetry.dev-dependencies]
 pylint = "^2.6.0"
 black = "^20.8b1"
 pytest = "^6.2.1"
 pytest-cov = "^2.11.0"
+mkdocs-material = "^7.0.2"
+flake8 = "^3.9.0"
+coverage = "^5.5"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.black]
+line-length = 120
  
+[tool.pytest.ini_options]
+minversion = "6.0"
+addopts = "-ra -q --cov=mealie"
+python_files = 'test_*'
+python_classes = '*Tests'
+python_functions = 'test_*'
+testpaths = [
+    "tests",
+]
+
+[tool.coverage.report]
+skip_empty = true
