@@ -1,9 +1,9 @@
 from typing import List
 
+from mealie.db.models.model_base import SqlAlchemyBase
+from pydantic import BaseModel
 from sqlalchemy.orm import load_only
 from sqlalchemy.orm.session import Session
-
-from db.sql.model_base import SqlAlchemyBase
 
 
 class BaseDocument:
@@ -11,21 +11,23 @@ class BaseDocument:
         self.primary_key: str
         self.store: str
         self.sql_model: SqlAlchemyBase
+        self.orm_mode = False
+        self.schema: BaseModel
 
     # TODO: Improve Get All Query Functionality
-    def get_all(
-        self, session: Session, limit: int = None, order_by: str = None
-    ) -> List[dict]:
-        list = [x.dict() for x in session.query(self.sql_model).limit(limit).all()]
+    def get_all(self, session: Session, limit: int = None, order_by: str = None) -> List[dict]:
 
-        if limit == 1:
-            return list[0]
+        if self.orm_mode:
+            return [self.schema.from_orm(x) for x in session.query(self.sql_model).limit(limit).all()]
 
-        return list
+        # list = [x.dict() for x in session.query(self.sql_model).limit(limit).all()]
 
-    def get_all_limit_columns(
-        self, session: Session, fields: List[str], limit: int = None
-    ) -> List[SqlAlchemyBase]:
+        # if limit == 1:
+        #     return list[0]
+
+        # return list
+
+    def get_all_limit_columns(self, session: Session, fields: List[str], limit: int = None) -> List[SqlAlchemyBase]:
         """Queries the database for the selected model. Restricts return responses to the
         keys specified under "fields"
 
@@ -37,11 +39,7 @@ class BaseDocument:
         Returns:
             list[SqlAlchemyBase]: Returns a list of ORM objects
         """
-        results = (
-            session.query(self.sql_model).options(load_only(*fields)).limit(limit).all()
-        )
-
-        return results
+        return session.query(self.sql_model).options(load_only(*fields)).limit(limit).all()
 
     def get_all_primary_keys(self, session: Session) -> List[str]:
         """Queries the database of the selected model and returns a list
@@ -53,15 +51,11 @@ class BaseDocument:
         Returns:
             list[str]:
         """
-        results = session.query(self.sql_model).options(
-            load_only(str(self.primary_key))
-        )
+        results = session.query(self.sql_model).options(load_only(str(self.primary_key)))
         results_as_dict = [x.dict() for x in results]
         return [x.get(self.primary_key) for x in results_as_dict]
 
-    def _query_one(
-        self, session: Session, match_value: str, match_key: str = None
-    ) -> SqlAlchemyBase:
+    def _query_one(self, session: Session, match_value: str, match_key: str = None) -> SqlAlchemyBase:
         """Query the sql database for one item an return the sql alchemy model
         object. If no match key is provided the primary_key attribute will be used.
 
@@ -72,18 +66,12 @@ class BaseDocument:
         Returns:
             Union[Session, SqlAlchemyBase]: Will return both the session and found model
         """
-        if match_key == None:
+        if match_key is None:
             match_key = self.primary_key
 
-        result = (
-            session.query(self.sql_model).filter_by(**{match_key: match_value}).one()
-        )
+        return session.query(self.sql_model).filter_by(**{match_key: match_value}).one()
 
-        return result
-
-    def get(
-        self, session: Session, match_value: str, match_key: str = None, limit=1
-    ) -> dict or List[dict]:
+    def get(self, session: Session, match_value: str, match_key: str = None, limit=1) -> BaseModel or List[BaseModel]:
         """Retrieves an entry from the database by matching a key/value pair. If no
         key is provided the class objects primary key will be used to match against.
 
@@ -96,23 +84,19 @@ class BaseDocument:
         Returns:
             dict or list[dict]:
         """
-        if match_key == None:
+        if match_key is None:
             match_key = self.primary_key
 
-        result = (
-            session.query(self.sql_model)
-            .filter_by(**{match_key: match_value})
-            .limit(limit)
-            .all()
-        )
-        db_entries = [x.dict() for x in result]
+        result = session.query(self.sql_model).filter_by(**{match_key: match_value}).limit(limit).all()
 
         if limit == 1:
-            return db_entries[0]
+            try:
+                return self.schema.from_orm(result[0])
+            except IndexError:
+                return None
+        return [self.schema.from_orm(x) for x in result]
 
-        return db_entries
-
-    def create(self, session: Session, document: dict) -> dict:
+    def create(self, session: Session, document: dict) -> BaseModel:
         """Creates a new database entry for the given SQL Alchemy Model.
 
         Args: \n
@@ -124,14 +108,12 @@ class BaseDocument:
         """
         new_document = self.sql_model(session=session, **document)
         session.add(new_document)
-        return_data = new_document.dict()
         session.commit()
 
-        return return_data
+        return self.schema.from_orm(new_document)
 
-    def update(self, session: Session, match_value: str, new_data: str) -> dict:
+    def update(self, session: Session, match_value: str, new_data: str) -> BaseModel:
         """Update a database entry.
-
         Args: \n
             session (Session): Database Session
             match_value (str): Match "key"
@@ -143,17 +125,12 @@ class BaseDocument:
 
         entry = self._query_one(session=session, match_value=match_value)
         entry.update(session=session, **new_data)
-        return_data = entry.dict()
-        session.commit()
 
-        return return_data
+        session.commit()
+        return self.schema.from_orm(entry)
 
     def delete(self, session: Session, primary_key_value) -> dict:
-        result = (
-            session.query(self.sql_model)
-            .filter_by(**{self.primary_key: primary_key_value})
-            .one()
-        )
+        result = session.query(self.sql_model).filter_by(**{self.primary_key: primary_key_value}).one()
 
         session.delete(result)
         session.commit()
