@@ -1,6 +1,6 @@
 import { baseURL } from "./api-utils";
 import { apiReq } from "./api-utils";
-import { store } from "../store/store";
+import { store } from "../store";
 import { router } from "../main";
 import qs from "qs";
 
@@ -18,7 +18,12 @@ const recipeURLs = {
   updateImage: slug => `${prefix}${slug}/image`,
 };
 
-export default {
+export const recipeAPI = {
+  /**
+   * Create a Recipe by URL
+   * @param {string} recipeURL
+   * @returns {string} Recipe Slug
+   */
   async createByURL(recipeURL) {
     let response = await apiReq.post(recipeURLs.createByURL, {
       url: recipeURL,
@@ -38,6 +43,7 @@ export default {
 
   async create(recipeData) {
     let response = await apiReq.post(recipeURLs.create, recipeData);
+    store.dispatch("requestRecentRecipes");
     return response.data;
   },
 
@@ -57,15 +63,13 @@ export default {
   },
 
   async update(data) {
-    const recipeSlug = data.slug;
-
-    let response = await apiReq.put(recipeURLs.update(recipeSlug), data);
+    let response = await apiReq.put(recipeURLs.update(data.slug), data);
     store.dispatch("requestRecentRecipes");
     return response.data;
   },
 
   async delete(recipeSlug) {
-    apiReq.delete(recipeURLs.delete(recipeSlug));
+    await apiReq.delete(recipeURLs.delete(recipeSlug));
     store.dispatch("requestRecentRecipes");
     router.push(`/`);
   },
