@@ -1,7 +1,7 @@
 import { apiReq } from "./api-utils";
 
-export default {
-  // import api from "@/api";
+export const utilsAPI = {
+  // import { api } from "@/api";
   async uploadFile(url, fileObject) {
     let response = await apiReq.post(url, fileObject, {
       headers: {
