@@ -20,7 +20,7 @@
 </template>
 
 <script>
-import api from "@/api";
+import { api } from "@/api";
 import utils from "@/utils";
 import MealPlanCard from "./MealPlanCard";
 export default {
