@@ -2,38 +2,56 @@
   <div class="text-center">
     <v-dialog v-model="addRecipe" width="650" @click:outside="reset">
       <v-card :loading="processing">
-        <v-card-title class="headline"
-          >{{ $t("new-recipe.from-url") }}
-        </v-card-title>
+        <v-app-bar dark color="primary mb-2">
+          <v-icon large left v-if="!processing">
+            mdi-link
+          </v-icon>
+          <v-progress-circular
+            v-else
+            indeterminate
+            color="white"
+            large
+            class="mr-2"
+          >
+          </v-progress-circular>
 
-        <v-card-text>
-          <v-form ref="urlForm">
+          <v-toolbar-title class="headline">
+            {{ $t("new-recipe.from-url") }}
+          </v-toolbar-title>
+
+          <v-spacer></v-spacer>
+        </v-app-bar>
+        <v-form ref="urlForm" @submit.prevent="createRecipe">
+          <v-card-text>
             <v-text-field
               v-model="recipeURL"
               :label="$t('new-recipe.recipe-url')"
               required
               validate-on-blur
               autofocus
+              class="mt-1"
               :rules="[isValidWebUrl]"
+              :hint="$t('new-recipe.url-form-hint')"
+              persistent-hint
             ></v-text-field>
-          </v-form>
 
-          <v-alert v-if="error" color="red" outlined type="success">
-            {{ $t("new-recipe.error-message") }}
-          </v-alert>
-        </v-card-text>
+            <v-alert v-if="error" color="red" outlined type="success">
+              {{ $t("new-recipe.error-message") }}
+            </v-alert>
+          </v-card-text>
 
-        <v-divider></v-divider>
+          <v-divider></v-divider>
 
-        <v-card-actions>
-          <v-spacer></v-spacer>
-          <v-btn color="grey" text @click="reset">
-            {{ $t("general.close") }}
-          </v-btn>
-          <v-btn color="success" text @click="createRecipe">
-            {{ $t("general.submit") }}
-          </v-btn>
-        </v-card-actions>
+          <v-card-actions>
+            <v-spacer></v-spacer>
+            <v-btn color="grey" text @click="reset">
+              {{ $t("general.close") }}
+            </v-btn>
+            <v-btn color="success" text type="submit" :loading="processing">
+              {{ $t("general.submit") }}
+            </v-btn>
+          </v-card-actions>
+        </v-form>
       </v-card>
     </v-dialog>
     <v-speed-dial v-model="fab" fixed right bottom open-on-hover>
@@ -53,7 +71,7 @@
 </template>
 
 <script>
-import api from "@/api";
+import { api } from "@/api";
 
 export default {
   data() {
