@@ -2,34 +2,44 @@
   <v-form ref="file">
     <input ref="uploader" class="d-none" type="file" @change="onFileChanged" />
     <v-btn :loading="isSelecting" @click="onButtonClick" color="accent" text>
-      <v-icon left> mdi-cloud-upload </v-icon>
-      {{ $t("general.upload") }}
+      <v-icon left> {{ icon }}</v-icon>
+      {{ text ? text : defaultText }}
     </v-btn>
   </v-form>
 </template>
 
 <script>
-import api from "@/api";
+const UPLOAD_EVENT = "uploaded";
+import { api } from "@/api";
 export default {
   props: {
     url: String,
+    text: { default: "Upload" },
+    icon: { default: "mdi-cloud-upload" },
+    fileName: { default: "archive" },
   },
   data: () => ({
     file: null,
     isSelecting: false,
   }),
 
+  computed: {
+    defaultText() {
+      return this.$t("general.upload");
+    },
+  },
+
   methods: {
     async upload() {
       if (this.file != null) {
         this.isSelecting = true;
         let formData = new FormData();
-        formData.append("archive", this.file);
+        formData.append(this.fileName, this.file);
 
         await api.utils.uploadFile(this.url, formData);
 
         this.isSelecting = false;
-        this.$emit("uploaded");
+        this.$emit(UPLOAD_EVENT);
       }
     },
     onButtonClick() {
