@@ -1,11 +1,13 @@
 import Vue from "vue";
 import App from "./App.vue";
 import vuetify from "./plugins/vuetify";
-import store from "./store/store";
+import store from "./store";
 import VueRouter from "vue-router";
 import { routes } from "./routes";
 import i18n from "./i18n";
 import FlashMessage from "@smartweb/vue-flash-message";
+import "@mdi/font/css/materialdesignicons.css";
+import "typeface-roboto/index.css";
 
 Vue.use(FlashMessage);
 Vue.config.productionTip = false;
