@@ -1,11 +1,11 @@
 <template>
-  <div>
+  <v-container>
     <CategorySidebar />
     <CardSection
-      v-if="showRecent"
-      title="Recent"
+      v-if="siteSettings.showRecent"
+      :title="$t('page.recent')"
       :recipes="recentRecipes"
-      :card-limit="showLimit"
+      :card-limit="siteSettings.cardsPerSection"
     />
     <CardSection
       :sortable="true"
@@ -13,15 +13,15 @@
       :key="section.name + section.position"
       :title="section.name"
       :recipes="section.recipes"
-      :card-limit="showLimit"
+      :card-limit="siteSettings.cardsPerSection"
       @sort="sortAZ(index)"
       @sort-recent="sortRecent(index)"
     />
-  </div>
+  </v-container>
 </template>
 
 <script>
-import api from "@/api";
+import { api } from "@/api";
 import CardSection from "../components/UI/CardSection";
 import CategorySidebar from "../components/UI/CategorySidebar";
 export default {
@@ -35,14 +35,9 @@ export default {
     };
   },
   computed: {
-    showRecent() {
-      return this.$store.getters.getShowRecent;
-    },
-    showLimit() {
-      return this.$store.getters.getShowLimit;
-    },
-    homeCategories() {
-      return this.$store.getters.getHomeCategories;
+    siteSettings() {
+      console.log(this.$store.getters.getSiteSettings);
+      return this.$store.getters.getSiteSettings;
     },
     recentRecipes() {
       let recipes = this.$store.getters.getRecentRecipes;
@@ -55,14 +50,16 @@ export default {
   },
   methods: {
     async buildPage() {
-      this.homeCategories.forEach(async element => {
+      await this.$store.dispatch("requestSiteSettings");
+      this.siteSettings.categories.forEach(async element => {
         let recipes = await this.getRecipeByCategory(element.slug);
-        recipes.position = element.position;
+        if (recipes.recipes.length < 0) recipes.recipes = [];
+        console.log(recipes);
         this.recipeByCategory.push(recipes);
       });
     },
     async getRecipeByCategory(category) {
-      return await api.categories.get_recipes_in_category(category);
+      return await api.categories.getRecipesInCategory(category);
     },
     getRecentRecipes() {
       this.$store.dispatch("requestRecentRecipes");
