@@ -1,5 +1,5 @@
 <template>
-  <div>
+  <v-container>
     <v-row justify="center">
       <v-col cols="1"> </v-col>
       <v-col>
@@ -30,12 +30,12 @@
         />
       </v-col>
     </v-row>
-  </div>
+  </v-container>
 </template>
 
 <script>
-import SearchBar from "../components/UI/SearchBar";
-import RecipeCard from "../components/UI/RecipeCard";
+import SearchBar from "../components/UI/Search/SearchBar";
+import RecipeCard from "../components/Recipe/RecipeCard";
 export default {
   components: {
     SearchBar,
