@@ -1,5 +1,6 @@
-import api from "@/api";
+import { api } from "@/api";
 import Vuetify from "../../plugins/vuetify";
+import axios from "axios";
 
 function inDarkMode(payload) {
   let isDark;
@@ -16,8 +17,11 @@ function inDarkMode(payload) {
 
 const state = {
   activeTheme: {},
-  darkMode: "system",
+  darkMode: "light",
   isDark: false,
+  isLoggedIn: false,
+  token: "",
+  userData: {},
 };
 
 const mutations = {
@@ -35,15 +39,47 @@ const mutations = {
       state.darkMode = payload;
     }
   },
+  setIsLoggedIn(state, payload) {
+    state.isLoggedIn = payload;
+  },
+  setToken(state, payload) {
+    state.isLoggedIn = true;
+    axios.defaults.headers.common["Authorization"] = `Bearer ${payload}`;
+    state.token = payload;
+  },
+
+  setUserData(state, payload) {
+    state.userData = payload;
+  },
 };
 
 const actions = {
+  async requestUserData({ commit }) {
+    const userData = await api.users.self();
+    commit("setUserData", userData);
+  },
+
   async resetTheme({ commit }) {
     const defaultTheme = await api.themes.requestByName("default");
     if (defaultTheme.colors) {
       Vuetify.framework.theme.themes.dark = defaultTheme.colors;
       Vuetify.framework.theme.themes.light = defaultTheme.colors;
       commit("setTheme", defaultTheme);
+    }
+  },
+
+  async refreshToken({ commit, getters }) {
+    if (!getters.getIsLoggedIn) {
+      commit("setIsLoggedIn", false); // This is to be here... for some reasons?  ¯\_(ツ)_/¯
+      console.log("Not Logged In");
+      return;
+    }
+    try {
+      let authResponse = await api.users.refresh();
+      commit("setToken", authResponse.access_token);
+    } catch {
+      console.log("Failed Token Refresh, Logging Out...");
+      commit("setIsLoggedIn", false);
     }
   },
 
@@ -63,6 +99,9 @@ const getters = {
   getActiveTheme: state => state.activeTheme,
   getDarkMode: state => state.darkMode,
   getIsDark: state => state.isDark,
+  getIsLoggedIn: state => state.isLoggedIn,
+  getToken: state => state.token,
+  getUserData: state => state.userData,
 };
 
 export default {
