@@ -1,15 +1,12 @@
-from app_config import SQLITE_FILE, USE_SQL
+from mealie.core.config import settings
 from sqlalchemy.orm.session import Session
 
-from db.sql.db_session import sql_global_init
+from mealie.db.models.db_session import sql_global_init
 
 sql_exists = True
 
-if USE_SQL:
-    sql_exists = SQLITE_FILE.is_file()
-    SessionLocal = sql_global_init(SQLITE_FILE)
-else:
-    raise Exception("Cannot identify database type")
+sql_exists = settings.SQLITE_FILE.is_file()
+SessionLocal = sql_global_init(settings.SQLITE_FILE)
 
 
 def create_session() -> Session:
