@@ -1,31 +1,29 @@
 import operator
 import shutil
 
-from app_config import BACKUP_DIR, TEMPLATE_DIR
-from db.db_setup import generate_session
 from fastapi import APIRouter, Depends, File, HTTPException, UploadFile
-from models.backup_models import BackupJob, ImportJob, Imports, LocalBackup
-from services.backups.exports import backup_all
-from services.backups.imports import ImportDatabase
+from mealie.core.config import app_dirs
+from mealie.db.db_setup import generate_session
+from mealie.routes.deps import get_current_user
+from mealie.schema.backup import BackupJob, ImportJob, Imports, LocalBackup
+from mealie.schema.snackbar import SnackResponse
+from mealie.services.backups import imports
+from mealie.services.backups.exports import backup_all
 from sqlalchemy.orm.session import Session
 from starlette.responses import FileResponse
-from utils.snackbar import SnackResponse
 
-router = APIRouter(prefix="/api/backups", tags=["Backups"])
+router = APIRouter(prefix="/api/backups", tags=["Backups"], dependencies=[Depends(get_current_user)])
 
 
 @router.get("/available", response_model=Imports)
 def available_imports():
     """Returns a list of avaiable .zip files for import into Mealie."""
     imports = []
-    templates = []
-    for archive in BACKUP_DIR.glob("*.zip"):
+    for archive in app_dirs.BACKUP_DIR.glob("*.zip"):
         backup = LocalBackup(name=archive.name, date=archive.stat().st_ctime)
         imports.append(backup)
 
-    for template in TEMPLATE_DIR.glob("*.*"):
-        templates.append(template.name)
-
+    templates = [template.name for template in app_dirs.TEMPLATE_DIR.glob("*.*")]
     imports.sort(key=operator.attrgetter("date"), reverse=True)
 
     return Imports(imports=imports, templates=templates)
@@ -40,7 +38,10 @@ def export_database(data: BackupJob, session: Session = Depends(generate_session
         templates=data.templates,
         export_recipes=data.options.recipes,
         export_settings=data.options.settings,
+        export_pages=data.options.pages,
         export_themes=data.options.themes,
+        export_users=data.options.users,
+        export_groups=data.options.groups,
     )
     try:
         return SnackResponse.success("Backup Created at " + export_path)
@@ -52,9 +53,9 @@ def export_database(data: BackupJob, session: Session = Depends(generate_session
 
 
 @router.post("/upload")
-def upload_backup_zipfile(archive: UploadFile = File(...)):
+def upload_backup_file(archive: UploadFile = File(...)):
     """ Upload a .zip File to later be imported into Mealie """
-    dest = BACKUP_DIR.joinpath(archive.filename)
+    dest = app_dirs.BACKUP_DIR.joinpath(archive.filename)
 
     with dest.open("wb") as buffer:
         shutil.copyfileobj(archive.file, buffer)
@@ -66,36 +67,32 @@ def upload_backup_zipfile(archive: UploadFile = File(...)):
 
 
 @router.get("/{file_name}/download")
-async def upload_nextcloud_zipfile(file_name: str):
+async def download_backup_file(file_name: str):
     """ Upload a .zip File to later be imported into Mealie """
-    file = BACKUP_DIR.joinpath(file_name)
+    file = app_dirs.BACKUP_DIR.joinpath(file_name)
 
     if file.is_file:
-        return FileResponse(
-            file, media_type="application/octet-stream", filename=file_name
-        )
+        return FileResponse(file, media_type="application/octet-stream", filename=file_name)
     else:
         return SnackResponse.error("No File Found")
 
 
 @router.post("/{file_name}/import", status_code=200)
-def import_database(
-    file_name: str, import_data: ImportJob, session: Session = Depends(generate_session)
-):
+def import_database(file_name: str, import_data: ImportJob, session: Session = Depends(generate_session)):
     """ Import a database backup file generated from Mealie. """
 
-    import_db = ImportDatabase(
+    return imports.import_database(
         session=session,
-        zip_archive=import_data.name,
+        archive=import_data.name,
         import_recipes=import_data.recipes,
+        import_settings=import_data.settings,
+        import_pages=import_data.pages,
+        import_themes=import_data.themes,
+        import_users=import_data.users,
+        import_groups=import_data.groups,
         force_import=import_data.force,
         rebase=import_data.rebase,
-        import_settings=import_data.settings,
-        import_themes=import_data.themes,
     )
-
-    imported = import_db.run()
-    return imported
 
 
 @router.delete("/{file_name}/delete", status_code=200)
@@ -103,7 +100,7 @@ def delete_backup(file_name: str):
     """ Removes a database backup from the file system """
 
     try:
-        BACKUP_DIR.joinpath(file_name).unlink()
+        app_dirs.BACKUP_DIR.joinpath(file_name).unlink()
     except:
         HTTPException(
             status_code=400,
