@@ -1,9 +1,9 @@
 from typing import List, Optional
 
-from db.database import db
-from db.db_setup import generate_session
 from fastapi import APIRouter, Depends, Query
-from models.recipe_models import AllRecipeRequest
+from mealie.db.database import db
+from mealie.db.db_setup import generate_session
+from mealie.schema.recipe import AllRecipeRequest
 from slugify import slugify
 from sqlalchemy.orm.session import Session
 
@@ -44,9 +44,7 @@ def get_all_recipes(
 
 
 @router.post("/api/recipes")
-def get_all_recipes_post(
-    body: AllRecipeRequest, session: Session = Depends(generate_session)
-):
+def get_all_recipes_post(body: AllRecipeRequest, session: Session = Depends(generate_session)):
     """
     Returns key data for all recipes based off the body data provided.
     For example, if slug, image, and name are provided you will recieve a list of
@@ -75,11 +73,9 @@ def get_all_recipes_post(
 @router.post("/api/recipes/category")
 def filter_by_category(categories: list, session: Session = Depends(generate_session)):
     """ pass a list of categories and get a list of recipes associated with those categories """
-    #! This should be refactored into a single database call, but I couldn't figure it out
-    in_category = [
-        db.categories.get(session, slugify(cat), limit=1) for cat in categories
-    ]
-    in_category = [cat.get("recipes") for cat in in_category]
+    # ! This should be refactored into a single database call, but I couldn't figure it out
+    in_category = [db.categories.get(session, slugify(cat), limit=1) for cat in categories]
+    in_category = [cat.recipes for cat in in_category if cat]
     in_category = [item for sublist in in_category for item in sublist]
     return in_category
 
@@ -87,8 +83,8 @@ def filter_by_category(categories: list, session: Session = Depends(generate_ses
 @router.post("/api/recipes/tag")
 async def filter_by_tags(tags: list, session: Session = Depends(generate_session)):
     """ pass a list of tags and get a list of recipes associated with those tags"""
-    #! This should be refactored into a single database call, but I couldn't figure it out
+    # ! This should be refactored into a single database call, but I couldn't figure it out
     in_tags = [db.tags.get(session, slugify(tag), limit=1) for tag in tags]
-    in_tags = [tag.get("recipes") for tag in in_tags]
+    in_tags = [tag.recipes for tag in in_tags]
     in_tags = [item for sublist in in_tags for item in sublist]
     return in_tags
