@@ -1,13 +1,13 @@
-from db.db_setup import generate_session
 from fastapi import APIRouter, Depends, File, Form, HTTPException
-from fastapi.logger import logger
 from fastapi.responses import FileResponse
-from models.recipe_models import RecipeURLIn
-from services.image_services import read_image, write_image
-from services.recipe_services import Recipe
-from services.scraper.scraper import create_from_url
+from mealie.db.database import db
+from mealie.db.db_setup import generate_session
+from mealie.routes.deps import get_current_user
+from mealie.schema.recipe import Recipe, RecipeURLIn
+from mealie.schema.snackbar import SnackResponse
+from mealie.services.image_services import read_image, write_image
+from mealie.services.scraper.scraper import create_from_url
 from sqlalchemy.orm.session import Session
-from utils.snackbar import SnackResponse
 
 router = APIRouter(
     prefix="/api/recipes",
@@ -16,53 +16,64 @@ router = APIRouter(
 
 
 @router.post("/create", status_code=201, response_model=str)
-def create_from_json(data: Recipe, db: Session = Depends(generate_session)) -> str:
+def create_from_json(
+    data: Recipe,
+    session: Session = Depends(generate_session),
+    current_user=Depends(get_current_user),
+) -> str:
     """ Takes in a JSON string and loads data into the database as a new entry"""
-    new_recipe_slug = data.save_to_db(db)
+    recipe: Recipe = db.recipes.create(session, data.dict())
 
-    return new_recipe_slug
+    return recipe.slug
 
 
 @router.post("/create-url", status_code=201, response_model=str)
-def parse_recipe_url(url: RecipeURLIn, db: Session = Depends(generate_session)):
+def parse_recipe_url(
+    url: RecipeURLIn,
+    session: Session = Depends(generate_session),
+    current_user=Depends(get_current_user),
+):
     """ Takes in a URL and attempts to scrape data and load it into the database """
 
     recipe = create_from_url(url.url)
-
-    recipe.save_to_db(db)
+    recipe: Recipe = db.recipes.create(session, recipe.dict())
 
     return recipe.slug
 
 
 @router.get("/{recipe_slug}", response_model=Recipe)
-def get_recipe(recipe_slug: str, db: Session = Depends(generate_session)):
+def get_recipe(recipe_slug: str, session: Session = Depends(generate_session)):
     """ Takes in a recipe slug, returns all data for a recipe """
-    recipe = Recipe.get_by_slug(db, recipe_slug)
 
-    return recipe
+    return db.recipes.get(session, recipe_slug)
 
 
 @router.put("/{recipe_slug}")
 def update_recipe(
-    recipe_slug: str, data: Recipe, db: Session = Depends(generate_session)
+    recipe_slug: str,
+    data: Recipe,
+    session: Session = Depends(generate_session),
+    current_user=Depends(get_current_user),
 ):
     """ Updates a recipe by existing slug and data. """
 
-    new_slug = data.update(db, recipe_slug)
+    recipe: Recipe = db.recipes.update(session, recipe_slug, data.dict())
 
-    return new_slug
+    return recipe.slug
 
 
 @router.delete("/{recipe_slug}")
-def delete_recipe(recipe_slug: str, db: Session = Depends(generate_session)):
+def delete_recipe(
+    recipe_slug: str,
+    session: Session = Depends(generate_session),
+    current_user=Depends(get_current_user),
+):
     """ Deletes a recipe by slug """
 
     try:
-        Recipe.delete(db, recipe_slug)
+        db.recipes.delete(session, recipe_slug)
     except:
-        raise HTTPException(
-            status_code=404, detail=SnackResponse.error("Unable to Delete Recipe")
-        )
+        raise HTTPException(status_code=404, detail=SnackResponse.error("Unable to Delete Recipe"))
 
     return SnackResponse.error(f"Recipe {recipe_slug} Deleted")
 
@@ -83,9 +94,10 @@ def update_recipe_image(
     image: bytes = File(...),
     extension: str = Form(...),
     session: Session = Depends(generate_session),
+    current_user=Depends(get_current_user),
 ):
     """ Removes an existing image and replaces it with the incoming file. """
     response = write_image(recipe_slug, image, extension)
-    Recipe.update_image(session, recipe_slug, extension)
+    db.recipes.update_image(session, recipe_slug, extension)
 
     return response
