@@ -1,23 +1,22 @@
-from db.database import db
-from db.db_setup import generate_session
 from fastapi import APIRouter, Depends
+from mealie.db.database import db
+from mealie.db.db_setup import generate_session
+from mealie.routes.deps import get_current_user
+from mealie.schema.snackbar import SnackResponse
 from sqlalchemy.orm.session import Session
-from utils.snackbar import SnackResponse
-
-from utils.snackbar import SnackResponse
 
 router = APIRouter(tags=["Recipes"])
 
 router = APIRouter(
-    prefix="/api/recipes/tags",
+    prefix="/api/tags",
     tags=["Recipe Tags"],
 )
 
 
-@router.get("/")
+@router.get("")
 async def get_all_recipe_tags(session: Session = Depends(generate_session)):
     """ Returns a list of available tags in the database """
-    return db.tags.get_all_primary_keys(session)
+    return db.tags.get_all_limit_columns(session, ["slug", "name"])
 
 
 @router.get("/{tag}")
@@ -27,7 +26,9 @@ def get_all_recipes_by_tag(tag: str, session: Session = Depends(generate_session
 
 
 @router.delete("/{tag}")
-async def delete_recipe_tag(tag: str, session: Session = Depends(generate_session)):
+async def delete_recipe_tag(
+    tag: str, session: Session = Depends(generate_session), current_user=Depends(get_current_user)
+):
     """Removes a recipe tag from the database. Deleting a
     tag does not impact a recipe. The tag will be removed
     from any recipes that contain it"""
