@@ -2,23 +2,23 @@ import shutil
 from pathlib import Path
 
 import requests
-from app_config import IMG_DIR
-from utils.logger import logger
+from fastapi.logger import logger
+from mealie.core.config import app_dirs
 
 
 def read_image(recipe_slug: str) -> Path:
-    if IMG_DIR.joinpath(recipe_slug).is_file():
-        return IMG_DIR.joinpath(recipe_slug)
-    else:
-        recipe_slug = recipe_slug.split(".")[0]
-        for file in IMG_DIR.glob(f"{recipe_slug}*"):
-            return file
+    if app_dirs.IMG_DIR.joinpath(recipe_slug).is_file():
+        return app_dirs.IMG_DIR.joinpath(recipe_slug)
+
+    recipe_slug = recipe_slug.split(".")[0]
+    for file in app_dirs.IMG_DIR.glob(f"{recipe_slug}*"):
+        return file
 
 
 def write_image(recipe_slug: str, file_data: bytes, extension: str) -> Path.name:
     delete_image(recipe_slug)
 
-    image_path = Path(IMG_DIR.joinpath(f"{recipe_slug}.{extension}"))
+    image_path = Path(app_dirs.IMG_DIR.joinpath(f"{recipe_slug}.{extension}"))
     with open(image_path, "ab") as f:
         f.write(file_data)
 
@@ -27,7 +27,7 @@ def write_image(recipe_slug: str, file_data: bytes, extension: str) -> Path.name
 
 def delete_image(recipe_slug: str) -> str:
     recipe_slug = recipe_slug.split(".")[0]
-    for file in IMG_DIR.glob(f"{recipe_slug}*"):
+    for file in app_dirs.IMG_DIR.glob(f"{recipe_slug}*"):
         return file.unlink()
 
 
@@ -44,7 +44,7 @@ def scrape_image(image_url: str, slug: str) -> Path:
                 image_url = image_url.get("url")
 
     filename = slug + "." + image_url.split(".")[-1]
-    filename = IMG_DIR.joinpath(filename)
+    filename = app_dirs.IMG_DIR.joinpath(filename)
 
     try:
         r = requests.get(image_url, stream=True)
