@@ -2,10 +2,12 @@ import shutil
 from pathlib import Path
 
 import yaml
-from app_config import IMG_DIR, TEMP_DIR
-from services.recipe_services import Recipe
+from fastapi.logger import logger
+from mealie.core.config import app_dirs
+from mealie.db.database import db
+from mealie.schema.recipe import Recipe
+from mealie.utils.unzip import unpack_zip
 from sqlalchemy.orm.session import Session
-from utils.unzip import unpack_zip
 
 try:
     from yaml import CLoader as Loader
@@ -49,42 +51,42 @@ def read_chowdown_file(recipe_file: Path) -> Recipe:
             "tags": recipe_data.get("tags").split(","),
         }
 
-        new_recipe = Recipe(**reformat_data)
+        reformated_list = [{"text": instruction} for instruction in reformat_data["recipeInstructions"]]
 
-        reformated_list = []
-        for instruction in new_recipe.recipeInstructions:
-            reformated_list.append({"text": instruction})
+        reformat_data["recipeInstructions"] = reformated_list
 
-        new_recipe.recipeInstructions = reformated_list
-
-        return new_recipe
+        return Recipe(**reformat_data)
 
 
 def chowdown_migrate(session: Session, zip_file: Path):
+
     temp_dir = unpack_zip(zip_file)
 
     with temp_dir as dir:
-        image_dir = TEMP_DIR.joinpath(dir, zip_file.stem, "images")
-        recipe_dir = TEMP_DIR.joinpath(dir, zip_file.stem, "_recipes")
+        chow_dir = next(Path(dir).iterdir())
+        image_dir = app_dirs.TEMP_DIR.joinpath(chow_dir, "images")
+        recipe_dir = app_dirs.TEMP_DIR.joinpath(chow_dir, "_recipes")
 
         failed_recipes = []
         successful_recipes = []
         for recipe in recipe_dir.glob("*.md"):
             try:
                 new_recipe = read_chowdown_file(recipe)
-                new_recipe.save_to_db(session)
-                successful_recipes.append(recipe.stem)
-            except:
+                db.recipes.create(session, new_recipe.dict())
+                successful_recipes.append(new_recipe.name)
+            except Exception as inst:
+                session.rollback()
+                logger.error(inst)
                 failed_recipes.append(recipe.stem)
 
         failed_images = []
         for image in image_dir.iterdir():
             try:
-                if not image.stem in failed_recipes:
-                    shutil.copy(image, IMG_DIR.joinpath(image.name))
-            except:
+                if image.stem not in failed_recipes:
+                    shutil.copy(image, app_dirs.IMG_DIR.joinpath(image.name))
+            except Exception as inst:
+                logger.error(inst)
                 failed_images.append(image.name)
-
         report = {"successful": successful_recipes, "failed": failed_recipes}
 
     return report
